+++ src/main/java/com/prgrms/board/global/common/SuccessMessage.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public enum SuccessMessage {
+
     GET_POSTS_SUCCESS("게시물 전체 조회 성공"),
     GET_POST_SUCCESS("게시물 조회 성공"),
     CREATE_POST_SUCCESS("게시물 생성 성공"),
